ParTree

Description

Author : Prabhat Totoo 2011

A binary leaf tree data structure with parallel operations.

Elements are placed on leaves and the inner nodes store the size of the subtree beneath it.

Type Definition

data Tree a

New data type for Tree with strictness annotations in the definition.

 data Tree a = Leaf a | Node !Int !(Tree a) !(Tree a)

For example, a tree representation for the list [1,2,3,4]

 myTree = (Node 4 (Node 2 (Leaf 1) (Leaf 2)) (Node 2 (Leaf 3) (Leaf 4)))

Constructors

Leaf a

Node !Int !(Tree a) !(Tree a)

Basics

size :: Tree a -> Int

Find the size of a tree.

link :: Tree a -> Tree a -> Tree a

Link two trees, one to the left and one to the right, making a node.

Parallel operations

All operations that follow are parallel implementations.

treeToList :: Tree a -> [a]

Convert a tree to list.

treeMap :: (a -> b) -> Tree a -> Tree b

Map a function f to each element of the tree in parallel, producing a new tree. Similar to map function in standard and parMap for its parallel equivalent.

treeFold :: NFData a => (a -> a -> a) -> a -> Tree a -> a

Parallel fold function on tree. The first function passed as argument needs to be both associative and commutative, e.g. sum and product operators. Both have type def Num a => a -> a -> a

treeElem :: Eq a => a -> Tree a -> Bool

Check if there is any occurrence of the element in the tree. Search proceeds by inspecting subtrees e.g. left and right branches in parallel.

treeReverse :: Tree a -> Tree a

Reverse all the elements in the tree.

treeMin :: Ord a => Tree a -> a

Return the smallest element from the tree.