RAList

Description

Author : Prabhat Totoo 2011

Alternative representation of Haskell's widely used list data structure based on Chris Okasaki's implementation of random-access list. This representation is intended to expose parallelism using a structure built around trees which is better suited for parallel processing.

Many basic list functions found in the Prelude and Data.List modules are provided. Qualify uses of these function names with an alias for this module.

Parallel functions are marked Par. and the time complexity as O(..).

Types

data RAList a

RAList is the new data type to be used instead of list. It is defined as a list of complete binary trees (implemented in the module ParTree).

 data RAList a = RAL !Int ![Digit a]
 data Digit a = Zero | One !(Tree a)
 data Tree a = Leaf a | Node !Int !(Tree a) !(Tree a) -- Defined in ParTree module.

Constructors

RAL !Int ![Digit a]

Basic operations

A number of basic operations on the new structure are implemented in order to support other main functions.

cons :: a -> RAList a -> RAList a

O(1) Similar to cons in list (:). It prepends an element to the RAList.

head :: RAList a -> a

O(1) Return the first element from the RAList.

tail :: RAList a -> RAList a

O(1) Return all elements, except the first.

empty :: RAList a

An empty RAList where the size is 0 hence indicating that the structure is empty.

 RAL 0 []

isEmpty :: RAList a -> Bool

O(1) Check if RAList is empty, that is, size s equals to 0 in:

 RAL s ts

length :: RAList a -> Int

O(1) Return the number of elements in the RAList. The corresponding function length in standard list has a time complexity of O(n).

Convertion operations

These are used for converting from/to standard list.

fromDataList :: [a] -> RAList a

Build a RAList from a list.

toDataList :: RAList a -> [a]

Convert a RAList to a list.

toDataList' :: NFData a => RAList a -> [a]

Par. An alternative (parallel) implementation of toDataList.

More advanced operations

which include lookup (search) by index, update, checking if element is in RAList (elem), obtaining sub-RAList from take, drop, filter and partition operations.

lookup :: Int -> RAList a -> a

O(log n) Lookup the RAList using the given index like (!!) in standard list. Index has to be in range, otherwise an error occurs.

update :: Int -> a -> RAList a -> RAList a

O(log n) Update the element at the given index with a new value.

elem :: Eq a => a -> RAList a -> Bool

Check for any occurrence of the given element in the structure. Seq implementation similar to elem from standard list.

pelem :: Eq a => a -> RAList a -> Bool

Par. parallel implementation of elem.

pelem' :: (Eq a, NFData a) => a -> RAList a -> Bool

Par. improved parallel version.

take :: Int -> RAList a -> RAList a

Take the first n elements.

drop :: Int -> RAList a -> RAList a

Drop the first n elements.

filter :: (a -> Bool) -> RAList a -> RAList a

Return a new RAList with only those elements that match the predicate.

partition :: (a -> Bool) -> RAList a -> (RAList a, RAList a)

Return 2 partitions of the given RAList. The first RAList in the tuple pair satisfies the predicate, while the second does not.

Map

RAList transformations.

ralMap :: (a -> b) -> RAList a -> RAList b

Sequential high-order map function which applies a function f to every element of the RAList.

 ralMap sqrt $ fromDataList [1,4,9]
 RAL 3 [One (Leaf 1.0),One (Node 2 (Leaf 2.0) (Leaf 3.0))]

pRalMap :: Strategy b -> (a -> b) -> RAList a -> RAList b

Par. parallel implementation of map function on RAList, using parRAList.

pRalMap' :: NFData b => (a -> b) -> RAList a -> RAList b

Par. another version that uses treeMap.

Fold

fold :: (a -> b -> a) -> a -> RAList b -> a

Sequential fold on the RAList structure, same type signature as foldl (folds element from the left).

 fold (+) 0 $ fromDataList [1..5]

parfold :: NFData a => (a -> a -> a) -> a -> RAList a -> a

Par. parallel implementaion of fold function. The collection of trees in the RAList is folded in parallel and at tree level, more parallelism is gained through the use of treeFold. A top level fold generates the final result. The function f has to be both associative and commutative.

parfold' :: NFData a => (a -> a -> a) -> a -> RAList a -> a

Par. an alternative parallel version of fold on RAList where the collection of trees are processed in order and parallelism is at sub-structure (tree level) only.

Parallel strategies

Strategies are defined for Haskell's list in the module Control.Parallel.Strategies. Following are probably the 2 mostly used strategies.

evalRAList :: Strategy a -> Strategy (RAList a)

Strategy combinator that walks over the RAList and applies the argument strategy s to every element. It generalises parRAList.

seqRAList is not given, but can be easily defined by

 seqRAList s = evalRAList (rseq `dot` s)

parRAList :: Strategy a -> Strategy (RAList a)

parRAList is obtained by composing the element strategy s with rpar.

 parRAList s = evalRAList (rpar `dot` s)

Example

 ralist = fromDataList [1..5]
 ralMap (+1) ralist `using` parRAList rdeepseq

Or

 pRalMap rdeepseq (+1) ralist

Benchmark applications

The following are implemented so that a direct comparison can be made with similar implementation in standard list found in the StdList module.

Sum

sum :: Num a => RAList a -> a

Sequential sum function that uses fold.

psum :: (NFData a, Num a) => RAList a -> a

Par. parallel sum function that uses parfold.

psum' :: (NFData a, Num a) => RAList a -> a

Par. parallel sum function that uses parfold'.

Factorial

facto :: Integer -> Integer

Sequential factorial function that uses fold.

pfacto :: Integer -> Integer

Par. parallel factorial function that uses parfold.

Sort

quicksort :: Ord a => RAList a -> RAList a

Sort a RAList of n random numbers using the quicksort algorithm.

Histogram

histo :: Int -> [Int] -> RAList Int

The histogram function counts the occurrences of each integer in a list of 5n integers chosen randomly from 0..n-1.

Minimum

pRalMin :: (Num a, Ord a) => RAList a -> a

Par. Return the smallest integer from the RAList.

pRalMin' :: (Num a, Ord a, NFData a) => RAList a -> a

Par. improved parallel version. min is commutative i.e. min x y = min y x, so it can be used in fold.

Nub

nub :: Eq a => RAList a -> [Digit a] -> RAList a

Par. The nub function removes duplicate elements from the RAList. Naive version.

nub1 :: Eq a => RAList a -> RAList a

nub-circular.

nub2 :: Eq a => RAList a -> RAList a

more advanced nub-circular. (Does not work).